backend/requirements.txt
fastapi>=0.111,<0.112
uvicorn[standard]>=0.30,<0.31
sqlalchemy>=2.0,<3
psycopg2-binary>=2.9,<3
pydantic>=2.7,<3
pydantic-settings>=2.4,<3
python-multipart>=0.0.9,<1
orjson>=3.9,<4
tiktoken>=0.7,<1
pinecone>=5.0.0

pymupdf>=1.24,<2
python-docx>=0.8.11,<1

sentence-transformers>=3.0.0
torch>=2.2.0

openai>=1.40,<2
tenacity>=8.3,<9
